+++ crdesigner/config/__init__.py
@@ -1,0 +1,3 @@
+import os
+
+CONFIG_DIR = os.path.dirname(__file__)