+++ crdesigner/map_conversion/osm2cr/converter_modules/graph_operations/road_graph/_graph.py
@@ -7,7 +7,7 @@ from typing import List, Set, Tuple, Optional
 from ordered_set import OrderedSet
 import numpy as np
 
-from crdesigner.map_conversion.osm2cr import config
+from crdesigner.config.osm_config import osm_config as config
 from crdesigner.map_conversion.common import geometry
 
 from ._graph_node import GraphNode