+++ crdesigner/map_conversion/osm2cr/converter_modules/intermediate_operations/traffic_light_generator.py
@@ -1,4 +1,4 @@
-from crdesigner.map_conversion.osm2cr import config
+from crdesigner.config.osm_config import osm_config as config
 from commonroad.scenario.traffic_sign import TrafficLight, TrafficLightCycleElement, TrafficLightState, \
     TrafficLightDirection