+++ crdesigner/ui/gui/controller/settings/settings_controller.py
@@ -1,118 +1,131 @@
-from PyQt5 import QtCore
-from PyQt5.QtWidgets import QMainWindow
+from PyQt5 import QtCore, QtWidgets
 
-import crdesigner.ui.gui.controller.settings.gui_settings_controller as gui
-import crdesigner.ui.gui.controller.settings.osm_settings_controller as osm
-import crdesigner.ui.gui.controller.settings.sumo_settings_controller as sumo
-from crdesigner.ui.gui.model.settings.gui_settings_model import gui_settings as gui_settings_model
-from crdesigner.ui.gui.utilities.gui_sumo_simulation import SUMO_AVAILABLE
+from crdesigner.config.gui_config import gui_config as gui_settings_model
+from crdesigner.config.settings_config import settings as settings_model, CONFIGS_TO_RENDER
+from crdesigner.ui.gui.utilities.draw_params_updater import set_draw_params
 from crdesigner.ui.gui.view.settings.settings_ui import SettingsUI
 
 
 class SettingsController:
+    """
+    Controller for the settings window. Mainly provides the logic for the buttons. And
+    connects the settings window to the settings models. Also sets up the UI.
+    """
 
     def __init__(self, parent):
-
+        """
+        Initialize the settings controller. As well as the settings UI.
+        """
         self.cr_designer = parent.mwindow_ui
         self.scenario_model = parent.scenario_model
         self.canvas = parent.animated_viewer_wrapper.cr_viewer.dynamic
-        self.settings_window = QMainWindow(None)
 
-        self.settings_ui = SettingsUI()
-        self.settings_ui.setupUi(self.settings_window, self.cr_designer)
+        self.settings_ui = SettingsUI(self.cr_designer)
 
-        self.gui_settings = gui.GUISettingController(self)
-        self.settings_ui.gui_settings = self.gui_settings.gui_settings_ui
-
-        self.osm_settings = osm.OSMSettingsController(self, None)
-        self.settings_ui.osm_settings = self.osm_settings.osm_settings_ui
-
-        self.sumo_settings = sumo.SUMOSettingsController(self, None)
-        self.settings_ui.sumo_settings = self.sumo_settings.sumo_settings_ui
+        for config in CONFIGS_TO_RENDER:
+            tab_name = config.__class__.__name__.replace("Config", "").upper()
+            self.settings_ui._add_tab(tab_name, config.LAYOUT)
 
         self.settings_ui.update_window()
-        self.settings_ui.retranslateUi(self.settings_window)
-        QtCore.QMetaObject.connectSlotsByName(self.settings_window)
-
-        if SUMO_AVAILABLE:
-            self.sumo_settings = sumo.SUMOSettingsController(self,
-                                                             config=parent.obstacle_toolbox.sumo_simulation.config)
+        self.settings_ui._retranslate_ui(self.settings_ui.settings)
+        QtCore.QMetaObject.connectSlotsByName(self.settings_ui.settings)
 
         self.connect_events()
-        self.update_ui_values()
-        self.settings_window.show()
 
     def connect_events(self):
-        """ connect buttons to callables """
-        # self.window.btn_restore_defaults.clicked.connect(self.restore_default_button)
+        """
+        Connect buttons to callables.
+        """
+        self.settings_ui.button_select_directory.clicked.connect(_select_directory)
         self.settings_ui.button_ok.clicked.connect(self.apply_close)
         self.settings_ui.button_cancel.clicked.connect(self.close)
-        self.settings_ui.button_settodefault.clicked.connect(self.clicked)
-        self.osm_settings.connect_events(self.settings_ui.osm_settings)
-        self.gui_settings.connect_events()
-        if SUMO_AVAILABLE:
-            self.sumo_settings.connect_events()
+        self.settings_ui.button_set_to_default.clicked.connect(_set_default)
 
     def close(self):
-        self.settings_window.close()
-        self.gui_settings.close()
-
-    def update_ui_values(self):
         """
-        sets the values of the settings window to the current values of config
+        Reset the settings to their previous values and close the settings window.
         """
-        self.gui_settings.update_ui_values()
-        if SUMO_AVAILABLE:
-            self.sumo_settings.update_ui_values()
-        self.osm_settings.update_ui_values()
-        self.cr_designer.update_window()
+        _restore_all_settings()
 
-    def save_to_config(self):
+        self._hide_and_update()
+
+    def apply_close(self):
         """
-        saves the values in the settings window to config.py
+        Save the settings to their respective files and close the settings window.
         """
-        self.gui_settings.save_to_config()
-        self.sumo_settings.save_to_config()
-        self.osm_settings.save_to_config()
+        if not _validate_all():
+            return
+        _save_all_to_yaml()
+        self._hide_and_update()
 
-    def has_valid_entries(self) -> bool:
+    def show(self):
         """
-        Check if the user input is valid. Otherwise, warn the user.
-
-        :return: bool whether the input is valid
+        Backup the current settings and show the settings window.
         """
-        # use if settings get extended
-        return True
+        _backup_settings()
+        self.settings_ui.settings.show()
 
-    def apply_close(self) -> None:
-        """
-        closes settings if settings could be saved to config
-        """
+    def _hide_and_update(self):
+        self._update_all_dependents()
+        self.settings_ui.settings.hide()
 
-        if self.has_valid_entries():
-            self.gui_settings.apply_close()
-            if SUMO_AVAILABLE:
-                self.sumo_settings.apply_close()
-            self.osm_settings.apply_close()
-            self.settings_window.close()
-            self.cr_designer.crdesigner_console_wrapper.text_browser.append("settings saved")
-            self.canvas.update_obstacle_trajectory_params()
-            if self.scenario_model.scenario_created():
-                self.cr_designer.animated_viewer_wrapper.cr_viewer.update_plot()
-            self.cr_designer.update_window()
+    def _update_all_dependents(self):
+        # GUI updates
+        set_draw_params(trajectory=gui_settings_model.DRAW_TRAJECTORY,
+                        intersection=gui_settings_model.DRAW_INTERSECTIONS,
+                        obstacle_label=gui_settings_model.DRAW_OBSTACLE_LABELS,
+                        obstacle_icon=gui_settings_model.DRAW_OBSTACLE_ICONS,
+                        obstacle_direction=gui_settings_model.DRAW_OBSTACLE_DIRECTION,
+                        obstacle_signal=gui_settings_model.DRAW_OBSTACLE_SIGNALS,
+                        occupancy=gui_settings_model.DRAW_OCCUPANCY,
+                        traffic_signs=gui_settings_model.DRAW_TRAFFIC_SIGNS,
+                        traffic_lights=gui_settings_model.DRAW_TRAFFIC_LIGHTS,
+                        incoming_lanelets=gui_settings_model.DRAW_INCOMING_LANELETS,
+                        successors=gui_settings_model.DRAW_SUCCESSORS,
+                        intersection_labels=gui_settings_model.DRAW_INTERSECTION_LABELS,
+                        colorscheme=self.cr_designer.colorscheme())
 
-        else:
-            print("invalid settings")
 
-    def clicked(self):
-        gui_settings_model.apply_set_to_default()
-        self.gui_settings.update_window()
+def _set_default():
+    """
+    Sets the default settings for all settings models.
+    """
+    for config in CONFIGS_TO_RENDER:
+        config.reset_settings()
 
-        self.osm_settings.apply_set_to_default()
-        if SUMO_AVAILABLE:
-            self.sumo_settings.restore_defaults()
-        self.settings_window.close()
-        self.cr_designer.crdesigner_console_wrapper.text_browser.append("default settings")
-        if self.scenario_model.scenario_created():
-            self.cr_designer.animated_viewer_wrapper.cr_viewer.update_plot()
-        self.cr_designer.update_window()
+
+def _select_directory():
+    """
+    Opens a file dialog to select a directory. If a directory is selected, the
+    CUSTOM_SETTINGS_DIR attribute of the settings model is set to the selected directory.
+    (This triggers the _load_custom_settings function in the settings model.)
+    """
+    directory = QtWidgets.QFileDialog.getExistingDirectory(caption="Select Directory")
+    if directory:
+        settings_model.CUSTOM_SETTINGS_DIR = directory
+        settings_model.notify_all()
+
+
+def _validate_all() -> bool:
+    valid = True
+    for config in CONFIGS_TO_RENDER:
+        if not config.validate_all_settings():
+            valid = False
+    return valid
+
+
+def _backup_settings():
+    for config in CONFIGS_TO_RENDER:
+        config.backup_settings()
+
+
+def _save_all_to_yaml():
+    for config in CONFIGS_TO_RENDER:
+        config.save_to_yaml_file()
+    settings_model.save_to_yaml_file()
+
+
+def _restore_all_settings():
+    for config in CONFIGS_TO_RENDER:
+        config.restore_settings()
+        config.notify_all()