+++ tutorials/conversion_examples/example_osm_to_commonroad.py
@@ -8,7 +8,7 @@ from commonroad.planning.planning_problem import PlanningProblemSet
 
 import crdesigner.map_conversion.osm2cr.converter_modules.converter as converter
 import crdesigner.map_conversion.osm2cr.converter_modules.cr_operations.export as ex
-from crdesigner.map_conversion.osm2cr import config
+from crdesigner.config.osm_config import osm_config as config
 from crdesigner.map_conversion.osm2cr.converter_modules.osm_operations.downloader import download_around_map
 
 # download a map